DecisionModel/requirements.txt
# DecisionModel 最小闭环依赖
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.20.0
sentence-transformers>=2.2.0
Pillow>=9.0.0
